Split Moment off into its own file
